move ReviewQueueFilterChips to its own folder
